+++ packages/react/tests/setup.ts
@@ -1,0 +1,1 @@
+import '@product-design-system/react/styles.css';